requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
bleach==3.1.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastparquet==0.3.3
future==0.18.2
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
hyperopt==0.2.3
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.31.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.2
notebook==6.0.3
numba==0.48.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
pylama==7.7.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
snowballstemmer==2.0.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thrift==0.13.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
